requirements.txt
pandas==1.3.0
googlesearch_python==1.0.1
PyMySQL==1.0.2
sshtunnel==0.4.0
requests==2.25.1
arxiv==1.4.2
beautifulsoup4==4.10.0
celery==4.4.3
redis==3.5.3
